requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
gunicorn==21.2.0
Flask-Login==0.6.3
Werkzeug==2.3.7
certifi==2024.8.30

# Dependências de teste
pytest==7.4.2
pytest-cov==4.1.0
pytest-html==3.2.0
pytest-mock==3.11.1
coverage==7.3.2
factory-boy==3.3.0
pymongo==4.6.0
faker==19.6.2
mongomock==4.1.2

# Dependências FastAPI / Migração
fastapi
uvicorn
motor
mangum
python-multipart
